requirements.txt
python-dotenv==1.1.1
google-genai==1.21.1
python-dotenv==1.1.1
google-cloud-documentai==3.5.0
tqdm==4.67.1
pandas==2.3.0
pandas-stubs==2.3.0.250703
gdown==5.2.0
